Move the testdata out to a separate .cpp so they can be reused
